v2.2.0 Движок обновлён до v3.4.0 Добавлена передача координат между игроками Добавлена генерация фона и передача его игрокам Сервер хранит координаты игроков Игрок смотрит в сторону мыши

+++ src/game/client/person/Stats.java
@@ -10,7 +10,7 @@ public class Stats {
     public double pFactorCrit, mFactorCrit;
     public int accuracy, evasion;
     public double speedCasting, speedReload;
-    public double speedRun;
+    public int speedRun;
     public int loadCapacity;
 
     public Stats(){
